+++ application/rendezvous/supprimerRdv.php
@@ -1,4 +1,12 @@
 <?php
+session_start();
+
+if (!isset($_SESSION['loggedin']) || $_SESSION['loggedin'] !== true) {
+    header('Location: ../index.php'); // Redirige vers la page de connexion
+    exit;
+}
+?>
+<?php
 require("../components/configDB.php");
 
 $requete = 'DELETE FROM rendez_vous WHERE id_rdv = ' . $_GET['id_rdv'];
